Create file for LegacyInactiveTabModel
