+++ Client/Frontend/Browser/Tabs/Legacy/LegacyInactiveTabs/LegacyInactiveTabViewModel.swift
@@ -14,54 +14,6 @@ protocol InactiveTabsCFRProtocol: AnyObject {
     func presentUndoSingleToast(completion: @escaping (Bool) -> Void)
 }
 
-enum InactiveTabStatus: String, Codable {
-    case normal
-    case inactive
-    case shouldBecomeInactive
-}
-
-struct InactiveTabStates: Codable {
-    var currentState: InactiveTabStatus?
-    var nextState: InactiveTabStatus?
-}
-
-struct LegacyInactiveTabModel: Codable {
-    // Contains [TabUUID String : InactiveTabState current or for next launch]
-    var tabWithStatus: [String: InactiveTabStates] = [String: InactiveTabStates]()
-
-    static let userDefaults = UserDefaults()
-
-    /// Check to see if we ever ran this feature before, this is mainly
-    /// to avoid tabs automatically going to their state on their first ever run
-    static var hasRunInactiveTabFeatureBefore: Bool {
-        get { return userDefaults.bool(forKey: PrefsKeys.KeyInactiveTabsFirstTimeRun) }
-        set(value) { userDefaults.setValue(value, forKey: PrefsKeys.KeyInactiveTabsFirstTimeRun) }
-    }
-
-    static func save(tabModel: LegacyInactiveTabModel) {
-        userDefaults.removeObject(forKey: PrefsKeys.KeyInactiveTabsModel)
-        let encoder = JSONEncoder()
-        if let encoded = try? encoder.encode(tabModel) {
-            userDefaults.set(encoded, forKey: PrefsKeys.KeyInactiveTabsModel)
-        }
-    }
-
-    static func get() -> LegacyInactiveTabModel? {
-        if let inactiveTabsModel = userDefaults.object(forKey: PrefsKeys.KeyInactiveTabsModel) as? Data {
-            do {
-                let jsonDecoder = JSONDecoder()
-                let inactiveTabModel = try jsonDecoder.decode(LegacyInactiveTabModel.self, from: inactiveTabsModel)
-                return inactiveTabModel
-            } catch {}
-        }
-        return nil
-    }
-
-    static func clear() {
-        userDefaults.removeObject(forKey: PrefsKeys.KeyInactiveTabsModel)
-    }
-}
-
 class LegacyInactiveTabViewModel {
     private var inactiveTabModel = LegacyInactiveTabModel()
     private var allTabs = [Tab]()
